feat: Implement FCM token registration and notification system
- Added FCM token registration and refresh endpoints in AuthController.
- Created FCMController for sending notifications to single and multiple devices.
- Introduced NotificationController for broadcasting notifications and retrieving user notifications.
- Implemented NotificationService to handle notification logic and FCM integration.
- Added FCMSettings configuration for FCM server key and sender ID.
- Updated AutoMapper profiles to include mappings for notification DTOs.
- Created FCMRequestDto for FCM token and device information.
- Added NotificationCreateDto and NotificationDto for notification data handling.
- Updated RepairRequestService to limit estimated duration for repair requests.
- Created IFCMService and INotificationService interfaces for service abstraction.
- Updated database context and entities to support notifications.

+++ AptCare.Repository/Entities/Issue.cs
@@ -28,7 +28,7 @@ namespace AptCare.Repository.Entities
 
         public bool IsEmergency { get; set; }
         public int RequiredTechnician { get; set; } 
-        public int EstimatedDuration { get; set; }
+        public double EstimatedDuration { get; set; }
         public ActiveStatus Status { get; set; }
 
         public ICollection<RepairRequest>? RepairRequests { get; set; }
